+++ Src/H.Necessaire/H.Necessaire.Runtime.Google.FirestoreDB/Core/Resources/DataBinGoogleFirestoreDbStorageResource.cs
@@ -391,7 +391,7 @@ namespace H.Necessaire.Runtime.Google.FirestoreDB.Core.Resources
 
             reporter.SetSourceInterval(interval);
         }
-        private async Task ReportProgress(string action, decimal progressValue)
+        private async Task ReportProgress(string action, double progressValue)
         {
             ProgressReporter reporter = ProgressReporter.Get(WellKnownFirestoreCallContextKey.FirestoreDataBinProgressionScopeID);
             if (reporter == null)
@@ -563,7 +563,7 @@ namespace H.Necessaire.Runtime.Google.FirestoreDB.Core.Resources
 
                 reporter.SetSourceInterval(interval);
             }
-            private void ReportProgress(string action, decimal progressValue)
+            private void ReportProgress(string action, double progressValue)
             {
                 ProgressReporter reporter = ProgressReporter.Get(WellKnownFirestoreCallContextKey.FirestoreDataBinProgressionScopeID);
                 if (reporter == null)
